+++ src/main/java/roomescape/service/ReservationService.java
@@ -3,9 +3,9 @@ package roomescape.service;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
+import roomescape.DTO.CreateReservationRequest;
 import roomescape.Repository.ReservationRepository;
 import roomescape.Repository.ReservationTimeRepository;
-import roomescape.domain.reservation.CreateReservationRequest;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservationtime.ReservationTime;
